// src/app/models/elprof.model.ts
export class Elprof {
    province: string = ''; //"省市別"
    countyCity: string = ''; //"縣市別"
    constituency: string = ''; //"選區別"
    townshipDistrict: string = ''; //"鄉鎮市區"
    village: string = ''; //"村里別"
    voteStation: string = ''; //"投開票所"
    validTicket: string = ''; //"有效票"
    invalidTicket: string = ''; //"無效票"
    voteNumber: string = ''; //"投票數"
    voterNumber: string = ''; //"選舉人數"
    population: string = ''; //"人口數"
    candidateTotal: string = ''; //"候選人數合計"
    electedTotal: string = ''; //"當選人數合計"
    candidateMen: string = ''; //"候選人數男"
    candidateWoman: string = ''; //"候選人數女"
    electedMen: string = ''; //"當選人數男"
    electedWoman: string = ''; //"當選人數女"
    voterNumberPopulation: string = ''; //"選舉人數對人口數"
    voteNumberVoterNumber: string = ''; //"投票數對選舉人數"
    electedTotalCandidateTotal: string = ''; //"當選人數對候選人數"
}