// src/app/models/elctks.model.ts
export class Elctks {
    province: string = ''; //"省市別"
    countyCity: string = ''; //"縣市別"
    constituency: string = ''; //"選區別"
    townshipDistrict: string = ''; //"鄉鎮市區"
    village: string = ''; //"村里別"
    voteStation: string = ''; //"投開票所"
    candidateNo: string = ''; //"候選人號次"
    voteNumber: string = ''; //"得票數"
    voteShare: string = ''; //"得票率"
    mark: string = ''; //"當選註記"
}