// src/app/models/tidy-cand.model.ts
export class TidyCand {
    politicalPartyNo: string = ''; //"政黨代號"
    politicalPartyName: string = ''; //"政黨名稱"
    candidateNo: string = ''; //"候選人號次"
    assistantName: string = ''; //"副手名字" 找該政黨politicalPartyNo一樣 assistant是Y的那個人
    name: string = ''; //"名字"
    mark: string = ''; //"當選註記"
}